PhotosList

+++ app/src/main/java/com/example/matchtrip/User.kt
@@ -6,4 +6,4 @@ import androidx.room.PrimaryKey
 @Entity
 data class User(val email: String, val password: String, val firstName: String, val lastName: String) {
     @PrimaryKey(autoGenerate = true)
-    var id: Int = 0 }
+    var id: Long = 0 }
